gfx で presentation 構築
wasm のビルドエラーを修正
