+++ dearx_gfx/src/lib.rs
@@ -1,12 +1,20 @@
-pub mod component;
 pub mod domain;
+pub mod experimental;
+
+#[cfg(not(target_arch = "wasm32"))]
+pub mod component;
+#[cfg(not(target_arch = "wasm32"))]
 mod renderer;
+#[cfg(not(target_arch = "wasm32"))]
 mod scene;
 
 #[cfg(feature = "viewer")]
+#[cfg(not(target_arch = "wasm32"))]
 pub mod viewer;
 
+#[cfg(not(target_arch = "wasm32"))]
 pub use renderer::Renderer;
+#[cfg(not(target_arch = "wasm32"))]
 pub use scene::{Scene, SceneUpdater};
 
 #[cfg(test)]
