+++ dearx_presentation/sycamore/Cargo.toml
@@ -6,6 +6,7 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+dearx_gfx = { path = "../../dearx_gfx" }
 serde-wasm-bindgen = "0.4.3"
 serde = { version = "1.0.140", features = ["derive"] }
 sycamore = "0.8"
